Extract Prompt Studio layout helpers
